[AI Chat]: Move AssociatedContentDelegate to its own file
