+++ components/ai_chat/core/browser/BUILD.gn
@@ -22,6 +22,8 @@ static_library("browser") {
     "ai_chat_tab_focus_metrics.h",
     "associated_archive_content.cc",
     "associated_archive_content.h",
+    "associated_content_delegate.cc",
+    "associated_content_delegate.h",
     "associated_content_driver.cc",
     "associated_content_driver.h",
     "associated_content_manager.cc",
