+++ components/ai_chat/core/browser/associated_archive_content.cc
@@ -5,17 +5,12 @@
 
 #include "brave/components/ai_chat/core/browser/associated_archive_content.h"
 
-#include <ios>
-#include <ostream>
 #include <string_view>
-#include <type_traits>
 #include <utility>
 
-#include "base/functional/callback.h"
 #include "base/logging.h"
-#include "base/memory/weak_ptr.h"
 #include "base/strings/utf_ostream_operators.h"
-#include "brave/components/ai_chat/core/browser/conversation_handler.h"
+#include "brave/components/ai_chat/core/browser/associated_content_delegate.h"
 
 namespace ai_chat {
 
@@ -59,8 +54,7 @@ std::u16string AssociatedArchiveContent::GetTitle() const {
   return title_;
 }
 
-void AssociatedArchiveContent::GetContent(
-    ConversationHandler::GetPageContentCallback callback) {
+void AssociatedArchiveContent::GetContent(GetPageContentCallback callback) {
   std::move(callback).Run(text_content_, is_video_, "");
 }
 
@@ -70,11 +64,6 @@ std::string_view AssociatedArchiveContent::GetCachedTextContent() const {
 
 bool AssociatedArchiveContent::GetCachedIsVideo() const {
   return is_video_;
-}
-
-base::WeakPtr<ConversationHandler::AssociatedContentDelegate>
-AssociatedArchiveContent::GetWeakPtr() {
-  return weak_ptr_factory_.GetWeakPtr();
 }
 
 }  // namespace ai_chat
