+++ components/ai_chat/core/browser/conversation_handler.h
@@ -7,13 +7,10 @@
 #define BRAVE_COMPONENTS_AI_CHAT_CORE_BROWSER_CONVERSATION_HANDLER_H_
 
 #include <cstdint>
-#include <map>
 #include <memory>
 #include <optional>
-#include <ostream>
 #include <string>
 #include <string_view>
-#include <tuple>
 #include <utility>
 #include <vector>
 
@@ -26,13 +23,12 @@
 #include "base/observer_list.h"
 #include "base/observer_list_types.h"
 #include "base/scoped_observation.h"
-#include "base/task/sequenced_task_runner.h"
 #include "base/types/expected.h"
 #include "brave/components/ai_chat/core/browser/ai_chat_credential_manager.h"
 #include "brave/components/ai_chat/core/browser/ai_chat_metrics.h"
+#include "brave/components/ai_chat/core/browser/associated_content_delegate.h"
 #include "brave/components/ai_chat/core/browser/engine/engine_consumer.h"
 #include "brave/components/ai_chat/core/browser/model_service.h"
-#include "brave/components/ai_chat/core/browser/text_embedder.h"
 #include "brave/components/ai_chat/core/browser/types.h"
 #include "brave/components/ai_chat/core/common/mojom/ai_chat.mojom-forward.h"
 #include "brave/components/ai_chat/core/common/mojom/ai_chat.mojom-shared.h"
@@ -40,7 +36,6 @@
 #include "mojo/public/cpp/bindings/pending_receiver.h"
 #include "mojo/public/cpp/bindings/receiver_set.h"
 #include "mojo/public/cpp/bindings/remote_set.h"
-#include "url/gurl.h"
 
 class AIChatUIBrowserTest;
 namespace mojo {
@@ -69,110 +64,8 @@ class ConversationHandler : public mojom::ConversationHandler,
                             public ModelService::Observer,
                             public ConversationHandlerForMetrics {
  public:
-  // |invalidation_token| is an optional parameter that will be passed back on
-  // the next call to |GetPageContent| so that the implementer may determine if
-  // the page content is static or if it needs to be fetched again. Most page
-  // content should be fetched again, but some pages are known to be static
-  // during their lifetime and may have expensive content fetching, e.g. videos
-  // with transcripts fetched over the network.
-  using GetPageContentCallback = base::OnceCallback<
-      void(std::string content, bool is_video, std::string invalidation_token)>;
   using GeneratedTextCallback =
       base::RepeatingCallback<void(const std::string& text)>;
-
-  // TODO(petemill): consider making SearchQuerySummary generic (StagedEntries)
-  // or a list of ConversationTurn objects.
-  using GetStagedEntriesCallback = base::OnceCallback<void(
-      const std::optional<std::vector<SearchQuerySummary>>& entries)>;
-
-  // TODO(https://github.com/brave/brave-browser/issues/45732): Move this to its
-  // own file and merge with AssociatedContentDriver.
-  // Supplements a conversation with associated page content
-  class AssociatedContentDelegate {
-   public:
-    class Observer : public base::CheckedObserver {
-     public:
-      ~Observer() override {}
-      virtual void OnNavigated(AssociatedContentDelegate* delegate) {}
-      virtual void OnTitleChanged(AssociatedContentDelegate* delegate) {}
-    };
-
-    AssociatedContentDelegate();
-    virtual ~AssociatedContentDelegate();
-
-    // Unique ID for the content. For browser Tab content, this should be
-    // a navigation ID that's re-used during back navigations.
-    virtual int GetContentId() const = 0;
-    // Get metadata about the current page
-    virtual GURL GetURL() const = 0;
-    virtual std::u16string GetTitle() const = 0;
-
-    // Implementer should fetch content from the "page" associated with this
-    // conversation.
-    // |is_video| lets the conversation know that the content is focused on
-    // video content so that various UI language can be adapted.
-    virtual void GetContent(GetPageContentCallback callback) = 0;
-    // Get current cache of content, if available. Do not perform any fresh
-    // fetch for the content.
-    virtual std::string_view GetCachedTextContent() const = 0;
-    virtual bool GetCachedIsVideo() const = 0;
-    // Get summarizer-key meta tag content from Brave Search SERP if exists and
-    // use it to fetch search query and summary from Brave search chatllm
-    // endpoint.
-    virtual void GetStagedEntriesFromContent(GetStagedEntriesCallback callback);
-    // Signifies whether the content has permission to open a conversation's UI
-    // within the browser.
-    virtual bool HasOpenAIChatPermission() const;
-    virtual void GetScreenshots(
-        mojom::ConversationHandler::GetScreenshotsCallback callback);
-
-    void GetTopSimilarityWithPromptTilContextLimit(
-        const std::string& prompt,
-        const std::string& text,
-        uint32_t context_limit,
-        TextEmbedder::TopSimilarityCallback callback);
-
-    void SetTextEmbedderForTesting(
-        std::unique_ptr<TextEmbedder, base::OnTaskRunnerDeleter>
-            text_embedder) {
-      text_embedder_ = std::move(text_embedder);
-    }
-    TextEmbedder* GetTextEmbedderForTesting() { return text_embedder_.get(); }
-
-    base::WeakPtr<AssociatedContentDelegate> GetWeakPtr() {
-      return weak_ptr_factory_.GetWeakPtr();
-    }
-
-    virtual void OnTitleChanged();
-
-    void AddObserver(Observer* observer);
-    void RemoveObserver(Observer* observer);
-
-    const std::string& uuid() const { return uuid_; }
-    void set_uuid(std::string uuid) { uuid_ = uuid; }
-
-   protected:
-    // Content has navigated
-    virtual void OnNewPage(int64_t navigation_id);
-
-   private:
-    void OnTextEmbedderInitialized(bool initialized);
-
-    std::string uuid_;
-    base::ObserverList<Observer> observers_;
-
-    // Owned by this class so that all associated conversation can benefit from
-    // a single cache as page content is unlikely to change between messages
-    // and conversations.
-    std::unique_ptr<TextEmbedder, base::OnTaskRunnerDeleter> text_embedder_;
-    std::vector<std::tuple<std::string,  // prompt
-                           std::string,  // text
-                           uint32_t,     // context_limit
-                           TextEmbedder::TopSimilarityCallback>>
-        pending_top_similarity_requests_;
-
-    base::WeakPtrFactory<AssociatedContentDelegate> weak_ptr_factory_{this};
-  };
 
   class Observer : public base::CheckedObserver {
    public:
