+++ ios/browser/api/ai_chat/associated_content_driver_ios.h
@@ -26,7 +26,7 @@ class AssociatedContentDriverIOS : public AssociatedContentDriver {
  protected:
   std::u16string GetPageTitle() const override;
   GURL GetPageURL() const override;
-  void GetPageContent(ConversationHandler::GetPageContentCallback callback,
+  void GetPageContent(GetPageContentCallback callback,
                       std::string_view invalidation_token) override;
   void GetSearchSummarizerKey(GetSearchSummarizerKeyCallback callback) override;
